requirements.txt
streamlit
google-generativeai>=0.5.2
supabase